Move annotations to package without JAX-RS naming
