+++ runtime/src/main/java/io/quarkiverse/zanzibar/annotations/FGAObject.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.zanzibar.jaxrs.annotations;
+package io.quarkiverse.zanzibar.annotations;
 
 import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.ElementType.TYPE;
@@ -11,10 +11,10 @@ import java.lang.annotation.Target;
 @Inherited
 @Target({ TYPE, METHOD })
 @Retention(RUNTIME)
-public @interface FGARequestObject {
-
-    String property();
+public @interface FGAObject {
 
     String type();
+
+    String id();
 
 }
