+++ runtime/src/main/java/io/quarkiverse/zanzibar/annotations/FGAPathObject.java
@@ -1,4 +1,4 @@
-package io.quarkiverse.zanzibar.jaxrs.annotations;
+package io.quarkiverse.zanzibar.annotations;
 
 import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.ElementType.TYPE;
@@ -11,7 +11,7 @@ import java.lang.annotation.Target;
 @Inherited
 @Target({ TYPE, METHOD })
 @Retention(RUNTIME)
-public @interface FGAQueryObject {
+public @interface FGAPathObject {
 
     String param();
